New map.  Also drafting some new map syntax.
Took a preliminary crack at teleporter syntax, and the two types of
level descriptions.  Teleporters aren't really that different from the
original game, and I'm open to changes.  Did not add the "win"
teleporter yet -- we need to sort out event triggering/listening, I
think.  We might also consider making it so that maps don't even
require win teleporters to be placed, as it is basically no-brainer
default behavior.

The new map is "Eulogy" from Not-AA, written out manually.  I changed
the name to Bodhi because a) I never really liked the original name, b)
I like this name, and c) holy hell this is such a clever name you guys.
 The word "Bodhi," like the level itself, shares a superficial likeness
to "Bwadi."