Numerous improvements to Celestials and Domes
Seemingly fixed an issue where all celestials could disappear when far
enough away from (0, 0, 0). In truth it probably would still occur if
you go far enough away, but I went very, VERY far (Stratocaster looked
like a screen smudge) and was unable to reproduce it.

I took the great work @jonathanvoss did with the new
Celestial/Starfield "flattening" and ran with it a bit, too. I did a
lot of experiments with differently scaled Celestials at different
radial sample levels, and concocted a pretty nice "LOD" sort-of
calculation for doing it dynamically.

I added the ability to scale Celestials up or down in the map XML, and
took the opportunity to normalize the size values for both Celestials
and Starfields. That is to say, a star with a size of 1.0 will look
like a particularly bright star, and anything higher than that will
tend to look a little on the silly side. A Celestial with a size of
30.0 will be the exact same size as they have been in the past.

I also added the ability to specify samples/planes for Domes in the map
XML.

Lastly, I uh moved the render.analyze() call to a more appropriate
location?